criando gerenciamento de obstaculos e sua movimentação / inicialização e alteração das structs nos tipos de dados de float para int e ja colocando a logica de liberacao de memoria

+++ include/default_structs.h
@@ -2,8 +2,8 @@
 #define DEFAULT_STRUCTS_H
 
 typedef struct {
-    float x;
-    float y;
+    int x;
+    int y;
 } Posicao;
 
 typedef struct {
@@ -13,8 +13,8 @@ typedef struct {
 } Animacao;
 
 typedef struct hitbox {
-    float largura;
-    float altura;
+    int largura;
+    int altura;
 } Hitbox;
 
 #endif /* DEFAULT_STRUCTS_H */
